Force plugin result to array when plugin is not Replaceable
A plugin can return an array or a scalar value. If the plugin
implements the ReplaceablePluginInterface interface the result can be
of any type, because the value associated with the plugin key gets
replaced. If the plugin does not implement the
ReplaceablePluginInterface interface its result will be merged with the
results of the other plugins having the same plugin key. In this last
case, if the result is not an array, it will be converted to one.